create general ResourceCoverageFields component
