--- src/components/resource/_fields/custom-coverage/custom-coverage-fields.css
@@ -1,5 +1,0 @@
-.coverage-fields-datepicker {
-  align-self: flex-start;
-  flex: 1 1 200px;
-  margin-right: 1em;
-}
