--- src/components/resource/_fields/custom-coverage/index.js
@@ -1,1 +1,0 @@
-export { default } from './custom-coverage-fields';
