--- src/components/resource/_fields/managed-coverage/index.js
@@ -1,1 +1,0 @@
-export { default } from './managed-coverage-fields';
