+++ src/components/resource/_fields/resource-coverage-fields/index.js
@@ -1,0 +1,1 @@
+export { default } from './resource-coverage-fields';
