+++ src/components/resource/_fields/resource-coverage-fields/resource-coverage-fields.css
@@ -4,8 +4,14 @@
   margin: 0.5em 1.75em var(--control-margin-bottom);
 }
 
-.coverage-fields-datepicker {
+.managed-coverage-fields-datepicker {
   align-self: flex-start;
   flex: 1 1 200px;
   margin: 0 0.5em;
+}
+
+.custom-coverage-fields-datepicker {
+  align-self: flex-start;
+  flex: 1 1 200px;
+  margin-right: 1em;
 }
